python-api/requirements.txt
ansi2html==1.6.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
docutils==0.17.1
htmlmin==0.1.12
idna==2.10
iniconfig==1.1.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-reporter==0.5.1
pytest-reporter-html1==0.8.1
pytest-reportlog==0.1.2
python-dotenv==0.17.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
